Fix dependencies
This patch edits three dependencies.

bitbucket.org/ww/autoneg
========================

This repository is not available any more. With this patch, this
repository uses a mirror in Github.
Moreover, the commit referenced in the Godep file does not seem to exist
in the mirror. I am changing the dependency to point to the latest
commit available in the mirror.

bitbucket.org/bertimus9/systemstat
==================================

Godep can only access Bitbucket through Bitbucket API version 1, which
is not available any more. I have created a mirror of that repository on
Github and this patch sets this repository to depend on the new mirror.

github.com/mholt/caddy/caddyfile
================================

The repository has moved to github.com/caddyserver/caddy. This patch
changes the import path to the new location.